+++ packages/core-gatsby-config/types/react-capcha.d.ts
@@ -1,0 +1,4 @@
+declare module 'react-google-recaptcha' {
+  const value: any;
+  export default value;
+}