+++ packages/mir-website/gatsby-node.ts
@@ -63,7 +63,7 @@ export const createPages: GatsbyNode['createPages'] = async ({ graphql, actions 
      * Automatically create pages based on the Page Collection in Contentful
      */
     const pageTemplate = path.resolve(`src/templates/page.template.tsx`);
-    pages.data.allContentfulPage.edges
+    pages?.data?.allContentfulPage.edges
       .filter((edge) => {
         if (!(edge && edge.node)) {
           return false;
@@ -111,22 +111,22 @@ export const createPages: GatsbyNode['createPages'] = async ({ graphql, actions 
     }
 
     const blogPostTemplate = path.resolve(`src/templates/blog-post.template.tsx`);
-    const blogPageRouteFR = pages.data.allContentfulPage.edges
+    const blogPageRouteFR = pages?.data?.allContentfulPage.edges
       .filter((edge) => edge.node.name.includes('Blogue') && edge.node.node_locale === 'fr-CA')
       .map((edge) => edge.node);
-    const blogPageRouteEN = pages.data.allContentfulPage.edges
+    const blogPageRouteEN = pages?.data?.allContentfulPage.edges
       .filter((edge) => edge.node.name.includes('Blogue') && edge.node.node_locale === 'en-CA')
       .map((edge) => edge.node);
 
-    log(`Creating blog posts under: ${blogPageRouteEN[0].route}`, {
+    log(`Creating blog posts under: ${blogPageRouteEN?.[0].route}`, {
       toolName: 'mir-website',
     });
 
-    log(`Creating blog posts under: ${blogPageRouteFR[0].route}`, {
+    log(`Creating blog posts under: ${blogPageRouteFR?.[0].route}`, {
       toolName: 'mir-website',
     });
 
-    blogPosts.data.allContentfulBlogPost.edges
+    blogPosts?.data?.allContentfulBlogPost.edges
       .filter((edge) => {
         if (!(edge && edge.node)) {
           return false;
@@ -136,12 +136,12 @@ export const createPages: GatsbyNode['createPages'] = async ({ graphql, actions 
       .forEach((edge, index) => {
         const path =
           edge.node.node_locale === 'fr-CA'
-            ? blogPageRouteFR[0]?.route
-              ? `${blogPageRouteFR[0]?.route}${edge.node.blogSlug}`
-              : `${blogPageRouteFR[0]?.route}`
-            : blogPageRouteEN[0]?.route
-            ? `${blogPageRouteEN[0]?.route}${edge.node.blogSlug}`
-            : `${blogPageRouteEN[0]?.route}`;
+            ? blogPageRouteFR?.[0]?.route
+              ? `${blogPageRouteFR?.[0]?.route}${edge.node.blogSlug}`
+              : `${blogPageRouteFR?.[0]?.route}`
+            : blogPageRouteEN?.[0]?.route
+            ? `${blogPageRouteEN?.[0]?.route}${edge.node.blogSlug}`
+            : `${blogPageRouteEN?.[0]?.route}`;
 
         log(`Creating blog post: ${path}`, {
           toolName: 'mir-website',