+++ packages/mir-website/src/components/info-section/info-section-type-7.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
-
 import { useStyles } from 'react-treat';
 import * as styleRefs from './info-section-type-7.treat';
-
 import { InfoTile } from './info-tile';
+import { SectionFragment } from 'types/graphql-types';
+import { SECTION_TYPE } from '../../templates/section.template';
 
 type OwnProps = SectionFragment;
 
@@ -13,14 +13,14 @@ export const InfoSectionType7: React.FC<OwnProps> = (props) => {
   return (
     <div className={`${styles.wrapper}`}>
       <div className={`${styles.container}`}>
-        {props.infoTiles.map((infoTile, index) => {
+        {props?.infoTiles?.map((infoTile, index) => {
           return (
             <InfoTile
               key={index}
-              variant="borderWhite"
-              title={infoTile.title}
-              illustration={infoTile.illustration}
-              text={infoTile.text}
+              variants={SECTION_TYPE.ACTION_PRIMARY}
+              title={infoTile?.title}
+              illustration={infoTile?.illustration}
+              text={infoTile?.text}
             ></InfoTile>
           );
         })}