+++ packages/mir-website/src/templates/blog.template.tsx
@@ -11,29 +11,31 @@ import { SectionFragment } from '../../types/graphql-types';
 export const BlogTemplate: React.FC<ProjectPageProps> = (props) => {
   const styles = useStyles(stylesRef);
 
-  const page = props.data.contentfulPage;
+  const page = props?.data?.contentfulPage;
 
   const bannerImagesSource = [
-    page.bannerImages?.medias[0]?.mobileFluidImage,
+    page?.bannerImages?.medias?.[0]?.mobileFluidImage,
     {
-      ...page.bannerImages?.medias[0]?.desktopFluidImage,
+      ...page?.bannerImages?.medias?.[0]?.desktopFluidImage,
       media: `(min-width: ${theme.layout.breakpoints.desktopSmall.px})`,
     },
   ];
 
-  const blogSection: SectionFragment[] = props.data.contentfulPage.sections.filter(
-    (section) => section.type.name === SECTION_TYPE.BLOG_PREVIEW
-  );
+  const blogSection: SectionFragment[] = props?.data?.contentfulPage?.sections?.filter(
+    (section) => section?.type?.name === SECTION_TYPE.BLOG_PREVIEW
+  ) as SectionFragment[];
 
   return (
     <main className={`${styles.wrapper}`}>
       <BannerSecondary imageData={bannerImagesSource} title={props.data?.contentfulPage?.bannerTitle}></BannerSecondary>
 
-      <BlogPreviewSection
-        pageRoute={props.data.contentfulPage.route}
-        variant="full"
-        {...blogSection[0]}
-      ></BlogPreviewSection>
+      {props?.data?.contentfulPage?.route && blogSection ? (
+        <BlogPreviewSection
+          pageRoute={props?.data?.contentfulPage?.route}
+          variant="full"
+          {...blogSection[0]}
+        ></BlogPreviewSection>
+      ) : null}
 
       {/* <Newsletter id="newsletter" section={actionSection1}></Newsletter> */}
     </main>