+++ packages/newrade-website/gatsby-config.js
@@ -41,17 +41,6 @@ exports.config = {
         core.getGatsbyReactSvgConfig(),
         core.getGastbyCorePluginConfig(),
         core.getGastbyPluginTreatConfig(),
-        {
-            resolve: `gatsby-source-graphql`,
-            options: {
-                typeName: `GraphCMS`,
-                fieldName: `gcms`,
-                url: process.env.GRAPH_CMS_API_URL_NEWRADE,
-                headers: {
-                    Authorization: `bearer ${process.env.GRAPH_CMS_AUTH_TOKEN_NEWRADE}`,
-                },
-            },
-        },
     ],
 };
 exports.default = exports.config;