+++ packages/newrade-website/types/mdx.d.ts
@@ -1,0 +1,6 @@
+declare module '@mdx-js/react' {
+  const value: any;
+  const MDXProvider: any;
+  export default value;
+  export { MDXProvider };
+}